salt: Add `certificates` pillar entry for calico
This pillar will be used by all nodes except if
the node role is only CA.

This pillar entry will be consumed by the beacon
and the reactor listening for certificate expiration
events.

If the path of an expired certificate matches one
in this list, the related `sls` will be run.

Refs: #1887